requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.10.2
idna==2.8
mysqlclient==1.4.2.post1
Pillow==5.4.1
PyMySQL==0.9.3
pytesseract==0.2.6
pytz==2019.1
requests==2.22.0
selenium==3.141.0
sqlparse==0.3.0
tencentcloud-sdk-python==3.0.80
urllib3==1.24.1
